+++ CHANGELOG.md
@@ -36,6 +36,11 @@
 
 ### Bug Fixes
 
+* Fix `attributes` false positives when a comment line appears between an attribute and a
+  variable declaration.  
+  [leno23](https://github.com/leno23)
+  [#4596](https://github.com/realm/SwiftLint/issues/4596)
+
 * Avoid false positives in `prefer_self_in_static_references` for generic
   constraints and generic parameter bounds such as `where A: P` and `<A: P>`
   in classes and extensions.